requirements.txt
fastapi
numpy
pillow
python-multipart
torch
transformers
uvicorn[standard]
